Remove duplicated 'User-Agent' header from object-updater's requests
When Object-Server saves async-update in pickle files,
'User-Agent: object-server %(pid)' (the key is title style) is included
in the pickles. However, Object-Updater will add
'user-agent: object-updater %(pid)' (the key is lower case style) to the
pickled headers and use it to make http connection. According to RFC
7230, each header field consists of a case-insensitive field name,
therefore either of 'User-Agent' and 'user-agent' should not be included
in the headers.

This patch removes old 'User-Agent' header from object-updater's
requests.

Change-Id: Ia624558395584457c718b311fe80e1a8406e22ad
Closes-Bug:#1635114