Optimize noop case for suffix rehash
REPLICATE calls where everything is in sync and no suffixes have been
invalidated are supposed to be pretty common and fairly cheap.  If an
invalidations files is empty there's no need to perform a truncation
write operation which will presumably at some point have to flush.

Everyone owes Pavel a quarter for the one billion less write ops in prod

... and Matt a nickel for 20ms of not sleep back every unittest run

As a drive by I remove a crufty confusing OSError exception handler
around an open call that would be using O_CREAT in a directory that it
either just created or opened a file from - it wasn't going to raise
ENOENT.  Similarly rather than loose sleep trying to reason about all
the crazy exceptions that actually *could* pop anywhere in this method,
instead I improve the logging where any such exception would be caught.
This way we can get the details we need to focus on only errors that
actually happen.

Author: Pavel Kvasnička <pavel.kvasnicka@firma.seznam.cz>
Co-Author: Matthew Oliver <matt@oliver.net.au>

Related-Change-Id: I64cadb1a3feb4d819d545137eecfc295389794f0

Change-Id: If712e4431322df5c3e84808ab2d815fd06c76426