Test current reclaim_age handling
This is just tests to demonstrate the current behavior - and provide
context of behavioral differences/consistency in the related change.

See the related change for context.

The related change slightly modifies the outcome of the
reclaim/REPLICATE test.

The related change has no effect on the config handling.

The related change greatly simplfies the reclaim_age option
cleanup_ondisk_files plumbing tests.

Related-Change-Id: I2b9189941ac29f6e3be69f76ff1c416315270916
Co-Author: Clay Gerrard <clay.gerrard@gmail.com>

Change-Id: I5b5f90bb898a335e6336f043710a05a44e3b810f