Add container sharding to Swift containers
NOTE: This is a rebase and push of the current code, 5 POCs worth of
potential cruft, things commented out, not quite tuned settings and knobs.
Really exactly what is in the HEAD of the POC. So alot of polishing is
required. I am currently working on the documentation, which I hope get
get up in the next few days.. at least the first version to help with
reviews and get y'all ready to fix my dodgy code!

This patch adds container sharding functionality to Swift's
containers. You can turn on sharding on a container by setting:

  X-Container-Sharding: on

A container-sharder daemon is added, which does all the heavy
lifting. This is to be run on all the container servers. It
scans the sharding marked containers and on each container it:

  - Audit's the container (This still needs to be reworked as things
    have changed since the last POC).
  - Deal with any misplaced objects.
  - Depending on the size of the container it will either:
    - Do nothing
    - Shard the container down into peices, or
    - attempt to merge the container into it's neighbour (again this
      needs to be double checked and debugged from last POC).

These have been quite a few changes to other areas of swift to make it
shard aware, the biggest are:

  - Introduction of Pivot_ranges table to the container database to
    track pivots.
  - PivotRanges class in utils.
  - Container backend can now hadle 2 databases when sharding, which
    introduces 3 states of container databases:
    - UNSHARDED
    - SHARDING
    - SHARDED

It is the SHARDING state in which the backend must deal with 2
containers. The old one will never we written to again.

This change is still a WIP. See the trello board to see what is left to
be done (https://trello.com/b/z6oKKI4Q/container-sharding) but in
general we need:

  - Tests
  - Shrink testing
  - Auditing
  - Documentation
  - General reviews, debugging and corrections.

Change-Id: Ice703dc6d98108ad251c43f824426d026e1f1d97