Correct ringbuilder's set_weight usage string
The large version of the set_weight usage is missing the
new weight value. This has lead to confusion on the mailing
list.

This patch adds a <new_weight> option to not only add the
missing argument but also to make the command more clear
by differentiate between the search weight (--weight) and
how to set the new.

Change-Id: I7a4c5a6a4438f647fa33d65e22d1863c27d345be
Closes-Bug: #1660214