Use real mocking in unit tests test_cname_lookup
Replacing function without mocking creates weird behaviors as the results
of the tests sometimes depends on the order of the execution. Use real
mocking to avoid such behaviors.

Example: using real mock showed that test test_host_is_storage_domain was
wrong because it was not "reseting" the lookup_cname function at the
beginning of the test.

Change-Id: Ibc60697c9c850e6e3a108f0d906e8906c3f53ced