remove remote qualifier from release notes branch scanning
Reno will automatically add the origin remote prefix to the branch name
if it does not find a local branch with the same name, so the input
instructions don't need to include "origin/". Removing them allows a
local developer to experiment with changes on local branches, and should
have no impact on the CI system build results.

Depends-On: Id30888b77f46b1f163c0a8175033b96aa3f9bff8
Change-Id: I819fa75d55f96320a24fc4182e5f1cb63c678d16
Signed-off-by: Doug Hellmann <doug@doughellmann.com>