Add OpenSuse SAIO build instructions
This patch adds OpenSuse to the build a SAIO development page.

OpenSuse's libssl.so naming is different then other Linux distros
and as such it can't simply use pip's cryptography wheel/binary,
which by default is linked to libssl.so.10.
To fix this, --no-binary cryptography was added to pip install:

  pip install --no-binary cryptography -r requirements.txt

Which forces the cryptography module's binding to be compiled
against the correct libssl.so library.

Change-Id: I6a070f33d670edbb887433530c44e2cb509f0c58