Make X-Backend-Replication consistent for HEAD
Current an X-Backend-Replication GET request for an expired object will
still open the expired datafile and return the object, but HEAD request
with the same headers will 404.  This can lead to some bad assumptions
in probetests and other places where we make direct HEAD requests

N.B. Because SSYNC replication does not make any HEAD requests this
change is immaterial to the correctness of the consistency engine.

Related-Change-Id: I7f90b732c3268cb852b64f17555c631d668044a8
Change-Id: Idc01970b37d1b77e1d48f9c4f4979f63ee771093