sharding: fix and tests for container GET all items
Fix container GET with an items=all query string parameter to return
all items even when there are no other params specified. Previously it
only had effect when another param such as marker or prefix was
included, but there appears to be a case in the container controller
where those params may not be present.

Add unit tests for GETs with items=all.

Change-Id: I446050d17286a0067b2f535774c3c3a569b36811