Merge Shard-{Account, Container} to 1 metadata
Currently in each shard, we store a pointer to the root container in:

  X-Container-Sysmeta-Shard-Account: <account>
  X-Container-Sysmeta-Shard-Container: <container>

There is no reason to store this as pieces of metadata, especially seeing
as they both must exist, that's extra checking.

This patch merges them into 1 single piece of metadata:

  X-Container-Sysmeta-Shard-Root: <account>/<container>

Further, the Backend-Shard-{Account, Container} for the container
updates/object updater has also been changed to Shard-Root.

Change-Id: Id01b20133e5b015a1e5650fbed7ec225890b565c