Make broker.build_shard_ranges filter results for marker etc
The broker already does the marker/end_marker filtering for
object listing so put similar smarts in the broker method to get a
list of shard ranges (i.e. relocate from container server GET
method). Curiously this was already documented :)

Also, rename the method to get_shard_ranges since really it is just
getting them from the db, and that avoids any suggestion that 'build'
might imply 'scan' or 'find'.

Fix for when obj is not in a shard range or when using obj and
reverse.

Add to unit test, in particular for reverse listing case.

Change-Id: I0f72ae76d3e782e857b8a30e674a562e48c24f72