Place update into sharded container
When the container server receives an object PUT or DELETE
and the container is sharded it will go look for the shard
to update and return a redirect.
If it fails to find a shard, then just accept it locally so
misplaced objects can deal with it later.

Currently we return a HTTPCreated which is clearly wrong as
the update hasn't actually be written anywhere.

Change-Id: I385f60780ace4f8861a3d3fa07c6f348fa1650a4