Eliminate _generate_object_list method
The method was overkill - we either passed it ShardRanges
or object rows, which made the method complex.

In the case of ShardRanges there is one call site (the other is in
a so far unused method) and it is simple to inline the conversion
to dicts.

Similarly with objects there is one caller, so inlining is simple.

The method was never called with the delete kwarg.

Change-Id: I74ab5bbce2fdc890117240fa6254a22ba7a04838