Include a shard's account in its name
One day we may want to use a different or additional account
to store shard containers and will be glad we included the
shard account in each shard range name.

Also encapsulate shard range name generation in ShardRange and
add ShardRange.create() as a convenience method to instantiate
a new ShardRange.

Change-Id: I86f7b1904f6af577cb59e807813a63860b6ec31d