Allow sharded container to be deleted (when empty)
This is a *simple* delete strategy - if the root container thinks it is
empty then it is OK to delete it.

The root will not consider itself empty until (a) all objects have
been deleted and (b) all shards have updated their usage (i.e. the
sharders have passed over them), so there may be some delay between
object deletes and the root being deletable. That is analogous to a
normal container being undeletable because some delete updates dropped
into an async pending queue.

We can do better, but this allows us to delete at some point.

Change-Id: I942b3b5a8295bb9957eddb15a149263a9037dd9a