Use the same hash and created_at value in shard db
When moving to sharding state, match the current hash and created_at of
the old locked db.

This patch also cleans up a few todos in the area, including placing
the tmp db in the tmp folder so cleanup is taken care of if something
dies leaving tmp dbs lying around.

Change-Id: I5196ecbd1a18472129fcc801f17eba8e8c5bc342