Fix suffix-byte-range responses for zero-byte replicated objects.
Previously, given a GET request like "Range: bytes=-12345" for a
zero-byte object, Swift would return a 206 response with the header
"Content-Range: bytes 0--1/0". This is clearly incorrect. Now Swift
returns a 200 with the whole (zero-byte) object.

Note: this does not fix the bug for EC objects, only for replicated
ones. The fix for EC objects will follow in a separate commit.

Change-Id: If1edb665b0ae000da78c4efff6faddd94d75da6b
Partial-Bug: 1736840