Re-introduce shard range updates to root after shard has sharded
...but don't update their stats when shrinking.

When cleaving shard ranges, their object stats may have changed since
being found, so the shard range stats are updated after cleaving. This
is an optimisation to get timely stats updates, but not strictly
necessary, because once replicated the cleaved shards will
independently start to update the root container.

However, when shrinking, the acceptor shard range should not be
updated with the cleaved broker stats because the cleaved broker will
only contain a subset of the existing acceptor's objects.

When a shard has been completely sharded, send updated shard ranges to
the root container - again, this is an optimisation to get timely
updates to the root.

Change-Id: Ic7d6ea0fea70caf2a37382164cbd460e8b5ec1bd