Set shard range state during sharding
Use shard range state to differentiate between shards that have been
found and had their container CREATED, and those that have also been
cleaved and are ACTIVE. This allows the root container to only
consider ACTIVE shard ranges when serving GET and HEAD requests.

Shard ranges cease to be ACTIVE when for example they are SHRUNK.

Listing a container while it is partially sharding now requires
the final part of the listing to be supplied by the sharding
container. This was apparently broken because that final
partial listing request sent '&end_marker=None' (yes, the
string None).

Change-Id: I436ce26742894b3e85560d4c007d386da55cdf29